pyproject.toml
[project]
name = "learinig-fastapi"
version = "0.1.0"
description = "mini-project to learn fastapi"
authors = [
    {name = "IvanKorolev"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
